add common project; move maestro credential to common
