+++ arcade-services.sln
@@ -118,6 +118,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Maestro.Authentication", "s
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProductConstructionService.Client", "src\ProductConstructionService\ProductConstructionService.Client\ProductConstructionService.Client.csproj", "{964FA796-358E-48AE-B75C-E42132600BCC}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Maestro.Common", "src\Maestro\Maestro.Common\Maestro.Common.csproj", "{16F086DD-8387-44BB-87D5-CD804355A110}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -524,6 +526,18 @@ Global
 		{964FA796-358E-48AE-B75C-E42132600BCC}.Release|x64.Build.0 = Release|Any CPU
 		{964FA796-358E-48AE-B75C-E42132600BCC}.Release|x86.ActiveCfg = Release|Any CPU
 		{964FA796-358E-48AE-B75C-E42132600BCC}.Release|x86.Build.0 = Release|Any CPU
+		{16F086DD-8387-44BB-87D5-CD804355A110}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{16F086DD-8387-44BB-87D5-CD804355A110}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{16F086DD-8387-44BB-87D5-CD804355A110}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{16F086DD-8387-44BB-87D5-CD804355A110}.Debug|x64.Build.0 = Debug|Any CPU
+		{16F086DD-8387-44BB-87D5-CD804355A110}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{16F086DD-8387-44BB-87D5-CD804355A110}.Debug|x86.Build.0 = Debug|Any CPU
+		{16F086DD-8387-44BB-87D5-CD804355A110}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{16F086DD-8387-44BB-87D5-CD804355A110}.Release|Any CPU.Build.0 = Release|Any CPU
+		{16F086DD-8387-44BB-87D5-CD804355A110}.Release|x64.ActiveCfg = Release|Any CPU
+		{16F086DD-8387-44BB-87D5-CD804355A110}.Release|x64.Build.0 = Release|Any CPU
+		{16F086DD-8387-44BB-87D5-CD804355A110}.Release|x86.ActiveCfg = Release|Any CPU
+		{16F086DD-8387-44BB-87D5-CD804355A110}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -566,6 +580,7 @@ Global
 		{BB8CC065-79D6-48CE-AD96-32EDDDC4D404} = {1A456CF0-C09A-4DE6-89CE-1110EED31180}
 		{70D48B7A-DBFE-4762-A83F-4617ECF80838} = {AE791E26-78E1-4936-BCF4-1BF5152CBBD6}
 		{964FA796-358E-48AE-B75C-E42132600BCC} = {243A4561-BF35-405A-AF12-AC57BB27796D}
+		{16F086DD-8387-44BB-87D5-CD804355A110} = {AE791E26-78E1-4936-BCF4-1BF5152CBBD6}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {32B9C883-432E-4FC8-A1BF-090EB033DD5B}
